consolidated task to annotated functions or const values and updated task running logic to follow those changes
