+++ geoflow-core/src/main/kotlin/jobs/SystemJob.kt
@@ -5,7 +5,7 @@ import tasks.SystemTask
 
 /**
  * Generic job for kjob to run a [SystemTask]. Requires a [runId], [pipelineRunTaskId] and [taskClassName] to find the
- * [SystemTask] subclass and instantiate the class, as well as a [runNext] flag to denote if the user asked to keep
+ * [SystemTask] function and class then function, as well as a [runNext] flag to denote if the user asked to keep
  * running tasks after the first one is completed. Subsequent tasks are only run when next task is a [SystemTask] and
  * the current task finished successfully
  *
