--- geoflow-core/src/main/kotlin/tasks/ConfirmRunInstance.kt
@@ -1,9 +1,0 @@
-package tasks
-
-/**
- * User task to validate that the user has confirmed details of run instance
- */
-class ConfirmRunInstance(pipelineRunTaskId: Long): UserTask(pipelineRunTaskId) {
-
-    override val taskId: Long = 1
-}
