+++ geoflow-core/src/main/kotlin/utils.kt
@@ -1,4 +1,3 @@
-import tasks.UserTask
 import java.time.*
 import java.time.format.DateTimeFormatter
 
@@ -11,14 +10,6 @@ fun formatInstantDefault(timestamp: Instant?) = timestamp
 fun formatInstantDateTime(timestamp: Instant?) = timestamp
     ?.atZone(ZoneId.systemDefault())
     ?.format(DateTimeFormatter.ISO_LOCAL_DATE_TIME) ?: ""
-
-fun getUserPipelineTask(pipelineRunTaskId: Long, taskClassName: String): UserTask {
-    return ClassLoader
-        .getSystemClassLoader()
-        .loadClass("tasks.$taskClassName")
-        .getConstructor(Long::class.java)
-        .newInstance(pipelineRunTaskId) as UserTask
-}
 
 inline fun <T> requireNotEmpty(collection: Collection<T>, lazyMessage: () -> Any) {
     if (collection.isEmpty()) {
